fix aws sdk async side effects in tests
