+++ packages/server/src/utils/S3.ts
@@ -1,21 +1,22 @@
 import AWS from 'aws-sdk';
 import { GetObjectOutput } from 'aws-sdk/clients/s3';
-import readline from 'readline';
 import { isDev } from '../lib/env';
-import { logError } from './logging';
 
-if (isDev) {
-    AWS.config.credentials = new AWS.SharedIniFileCredentials({
-        profile: 'membership',
-    });
-}
-const S3 = new AWS.S3();
+const getS3 = () => {
+    if (isDev) {
+        AWS.config.credentials = new AWS.SharedIniFileCredentials({
+            profile: 'membership',
+        });
+    }
+    return new AWS.S3();
+};
 
 export const fetchS3Data = (bucket: string, key: string): Promise<string> => {
-    return S3.getObject({
-        Bucket: bucket,
-        Key: key,
-    })
+    return getS3()
+        .getObject({
+            Bucket: bucket,
+            Key: key,
+        })
         .promise()
         .then((result: GetObjectOutput) => {
             if (result.Body) {
@@ -27,27 +28,4 @@ export const fetchS3Data = (bucket: string, key: string): Promise<string> => {
             }
         })
         .catch(err => Promise.reject(`Failed to fetch S3 object ${bucket}/${key}: ${err}`));
-};
-
-export const streamS3DataByLine = (
-    bucket: string,
-    key: string,
-    onLine: (line: string) => void,
-    onComplete?: () => void,
-): void => {
-    const input = S3.getObject({
-        Bucket: bucket,
-        Key: key,
-    }).createReadStream();
-
-    const stream = readline.createInterface({
-        input,
-        terminal: false,
-    });
-
-    stream.on('line', onLine);
-    if (onComplete) {
-        stream.on('close', onComplete);
-    }
-    stream.on('error', error => logError(`Error streaming from S3 for ${bucket}/${key}: ${error}`));
 };
